+++ common/src/main/java/dev/ftb/mods/ftbxmodcompat/ftbquests/jei/helper/JEIRecipeHelper.java
@@ -1,7 +1,6 @@
 package dev.ftb.mods.ftbxmodcompat.ftbquests.jei.helper;
 
 import dev.architectury.fluid.FluidStack;
-import dev.architectury.injectables.annotations.ExpectPlatform;
 import dev.ftb.mods.ftblibrary.config.Tristate;
 import dev.ftb.mods.ftblibrary.ui.input.Key;
 import dev.ftb.mods.ftbxmodcompat.FTBXModCompat;
@@ -10,11 +9,10 @@ import dev.ftb.mods.ftbxmodcompat.ftbquests.jei.LootCrateRecipeManagerPlugin;
 import dev.ftb.mods.ftbxmodcompat.ftbquests.jei.QuestRecipeManagerPlugin;
 import dev.ftb.mods.ftbxmodcompat.ftbquests.recipemod_common.BaseRecipeHelper;
 import mezz.jei.api.constants.VanillaTypes;
-import mezz.jei.api.ingredients.IIngredientTypeWithSubtypes;
+import mezz.jei.api.runtime.IBookmarkManager;
 import mezz.jei.api.runtime.IIngredientManager;
 import mezz.jei.api.runtime.IJeiRuntime;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.material.Fluid;
 
 import java.util.Collection;
 
@@ -29,16 +27,17 @@ public class JEIRecipeHelper extends BaseRecipeHelper {
         FTBQuestsJEIIntegration.showRecipes(fluid);
     }
 
-    @ExpectPlatform
-    public static IIngredientTypeWithSubtypes<Fluid,?> getNativeFluidType(FluidStack fluid) {
-        throw new AssertionError();
-    }
-
     @Override
     public Tristate toggleBookmark(ItemStack stack) {
         IJeiRuntime runtime = FTBQuestsJEIIntegration.runtime;
         if (runtime != null && !FTBXModCompat.isTMRVLoaded) {
-            return JEIInternalsHelper.toggleBookmark(stack, runtime);
+            return runtime.getIngredientManager().createTypedIngredient(stack, true).map(ingredient -> {
+                IBookmarkManager bookmarkManager = runtime.getBookmarkManager();
+                if (bookmarkManager.contains(ingredient)) {
+                    return bookmarkManager.remove(ingredient) ? Tristate.FALSE : Tristate.DEFAULT;
+                }
+                return bookmarkManager.add(ingredient) ? Tristate.TRUE : Tristate.DEFAULT;
+            }).orElse(Tristate.DEFAULT);
         }
         return Tristate.DEFAULT;
     }
@@ -48,10 +47,12 @@ public class JEIRecipeHelper extends BaseRecipeHelper {
         return "JEI";
     }
 
+    @Override
     protected void refreshQuests() {
         QuestRecipeManagerPlugin.INSTANCE.refresh();
     }
 
+    @Override
     protected void refreshLootcrates() {
         LootCrateRecipeManagerPlugin.INSTANCE.refresh();
     }
@@ -72,7 +73,8 @@ public class JEIRecipeHelper extends BaseRecipeHelper {
 
     @Override
     public boolean isBookmarkKey(Key key) {
-        // TODO non-API usage!  API only exposes the 'R' and 'U' mappings
-        return !FTBXModCompat.isTMRVLoaded && JEIInternalsHelper.isBookmarkK(key);
+        IJeiRuntime runtime = FTBQuestsJEIIntegration.runtime;
+        return runtime != null && !FTBXModCompat.isTMRVLoaded
+                && runtime.getKeyMappings().getBookmark().isActiveAndMatches(key.getInputMapping());
     }
 }